requirements.txt
django==5.1
Pillow==10.4.0
django-filter==24.3
django-extensions==3.2.3